pyproject.toml
[tool.poetry]
name = "fs-task"
version = "0.1.0"
description = ""
authors = ["Artsiom_PH <workartpharm@gmail.com>"]
readme = "README.md"
packages = [{include = "fs_task"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.3"
djangorestframework = "^3.14.0"
celery = "^5.2.7"
requests = "^2.28.1"
redis = "^4.4.0"
rabbitmq = "^0.2.0"
black = "^22.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
